Split flask_api.py into routes.py (for routing related functions) and config.py (for configuration functions). Moved flask_api.py init functionality into init file. Updated test_flask_api to reflect these changes.
