+++ knowledge_graph/__init__.py
@@ -1,0 +1,5 @@
+from flask import Flask
+
+app = Flask(__name__)
+
+from knowledge_graph import routes
