+++ knowledge_graph/test_flask_api.py
@@ -3,16 +3,16 @@ import tempfile
 
 import pytest
 
-from knowledge_graph import flask_api
+from knowledge_graph import routes
 
 #todo: finish this test...
 @pytest.fixture
 def client():
-    db_fd, flask_api.app.config['DATABASE'] = tempfile.mkstemp()
-    flask_api.app.config['TESTING'] = True
+    db_fd, routes.app.config['DATABASE'] = tempfile.mkstemp()
+    routes.app.config['TESTING'] = True
 
-    with flask_api.app.test_client() as client:
+    with routes.app.test_client() as client:
         yield client
 
     os.close(db_fd)
-    os.unlink(flask_api.app.config['DATABASE'])
+    os.unlink(routes.app.config['DATABASE'])
